Refonte des salons éphémères

+++ Outils/VoiceEphem/CheckState.py
@@ -1,36 +1,38 @@
-from Core.Fonctions.Embeds import embedAssertClassic
+from Outils.VoiceEphem.FormatagePattern import formatageVoiceEphem
 from Stats.SQL.ConnectSQL import connectSQL
-
+# ID Hub, ID Salon, ID Owner, Locked
 async def connectionVoiceEphem(guildOT,before,after,member):
     try:
         if after.channel!=None:
-            assert after.channel.id in guildOT.voicehub
+            hub=after.channel.id
+            assert hub in guildOT.voicehub
             connexion,curseur=connectSQL(member.guild.id,"VoiceEphem","Guild",None,None)
-            chan=curseur.execute("SELECT * FROM salons WHERE ID=0 ORDER BY Nombre ASC").fetchone()
-            if chan==None:
-                await member.send(embed=embedAssertClassic("Désolé, il n'y a plus de salon éphémère disponible sur le serveur {0}. Veuillez réessayer plus tard.".format(member.guild.name)))
+            nb=curseur.execute("SELECT Count() AS Count FROM salons WHERE IDHub={0}".format(hub)).fetchone()
+            if guildOT.voicehub[hub].limite==0:
+                limite=None
             else:
-                if guildOT.voicehub[after.channel.id]==0:
-                    limite=None
-                else:
-                    limite=guildOT.voicehub[after.channel.id]
-                voice=await member.guild.create_voice_channel(chan["Nom"],overwrites=after.channel.overwrites,category=after.channel.category,bitrate=after.channel.bitrate,user_limit=limite)
-                await member.move_to(voice)
-                curseur.execute("UPDATE salons SET ID={0} WHERE Nombre={1}".format(voice.id,chan["Nombre"]))
-                guildOT.voiceephem.append(voice.id)
-                connexion.commit()
+                limite=guildOT.voicehub[hub].limite
+            voice=await member.guild.create_voice_channel(formatageVoiceEphem(guildOT.voicehub[hub].pattern,member,nb["Count"]+1),overwrites=after.channel.overwrites,category=after.channel.category,bitrate=after.channel.bitrate,user_limit=limite)
+            await member.move_to(voice)
+            print(hub,voice.id,member.id)
+            curseur.execute("INSERT INTO salons VALUES ({0},{1},{2},False)".format(hub,voice.id,member.id))
+            guildOT.voiceephem.append(voice.id)
+            connexion.commit()
     except AssertionError:
         pass
     
     try:
         if before.channel!=None:
             assert before.channel.id in guildOT.voiceephem
-            assert len(before.channel.voice_states)==0
             connexion,curseur=connectSQL(member.guild.id,"VoiceEphem","Guild",None,None)
-            guildOT.voiceephem.remove(before.channel.id)
-            curseur.execute("UPDATE salons SET ID=0 WHERE ID={0}".format(before.channel.id))
-            connexion.commit()
-            await before.channel.delete()
+            if len(before.channel.voice_states)==0:
+                guildOT.voiceephem.remove(before.channel.id)
+                curseur.execute("DELETE FROM salons WHERE IDChannel={0}".format(before.channel.id))
+                connexion.commit()
+                await before.channel.delete()
+            elif curseur.execute("SELECT * FROM salons WHERE IDChannel={0}".format(before.channel.id)).fetchone()["IDOwner"]==member.id:
+                curseur.execute("UPDATE salons SET IDOwner={0} WHERE IDChannel={1}".format(before.channel.voice_states[0],before.channel.id))
+                connexion.commit()
     except AssertionError:
         pass
 
@@ -38,11 +40,12 @@ async def checkAll(guildOT,bot):
     if guildOT.voiceephem!=[]:
         connexion,curseur=connectSQL(guildOT.id,"VoiceEphem","Guild",None,None)
         for i in guildOT.voiceephem:
-            chan=bot.get_channel(i)
-            assert len(chan.voice_states)==0
-            guildOT.voiceephem.remove(chan.id)
-            curseur.execute("UPDATE salons SET ID=0 WHERE ID={0}".format(chan.id))
-            await chan.delete()
+            try:
+                chan=bot.get_channel(i)
+                assert len(chan.voice_states)==0
+                guildOT.voiceephem.remove(chan.id)
+                curseur.execute("DELETE FROM salons WHERE IDChannel={0}".format(chan.id))
+                await chan.delete()
+            except AssertionError:
+                pass
         connexion.commit()
-
-
